NovaBot_Pro/requirements.txt
python-telegram-bot[all]==20.6
yt-dlp==2024.4.9
openai==1.10.0
tinydb==4.8.0
apscheduler==3.10.4
flask==3.0.0
pillow==10.0.0